feat: refacto client.ts
