+++ server/src/services/bal.service.ts
@@ -1,7 +1,23 @@
+import axios from "axios"
+import http from "http"
+import https from "https"
+import { setupCache } from "axios-cache-interceptor"
 import config from "../config.js"
 import { ApiError, apiRateLimiter } from "../common/utils/apiUtils.js"
 import { sentryCaptureException } from "../common/utils/sentryUtils.js"
-import getApiClient from "../common/client.js"
+
+const getApiClient = (options) =>
+  setupCache(
+    axios.create({
+      timeout: 5000,
+      httpAgent: new http.Agent({ keepAlive: true }),
+      httpsAgent: new https.Agent({ keepAlive: true }),
+      ...options,
+    }),
+    {
+      ttl: 1000 * 60 * 10, // 10 Minutes
+    }
+  )
 
 interface IBALResponse {
   is_valid: boolean
